BOGUE: bin/actu.servicenow.sh: suppression de la base
# On supprimait systématiquement la base, même en mode sans --init.
- Et d'ailleurs même en mode --init ce n'est pas sympa (il peut y avoir d'autres données que celles ServiceNow dans la base, ne serait-ce que du paramétrage), donc on s'ajoute une belle note pour plus tard.

darcs-hash:121e0edaeefdc5dd38e773824d681bf999fd8092

+++ bin/actu.servicenow.sh
@@ -29,7 +29,7 @@ _actu()
 	_telecache "$csv" 120 $opttelech || return 1
 	
 	titre "Conversion CSV -> SQL"
-	rm -f "$BDD"
+	[ -n "$temp" ] || rm -f "$BDD" # /!\ On bute la base en mode --init. Pas sympa si d'autres sources que nous ont alimenté. # À FAIRE: un simple delete de tout ce qui est à nous.
 	php "$SCRIPTS/../app/maj.php"
 	php "$SCRIPTS/servicenowactu.php" "$csv" > "$sql"
